Add code review changes
