+++ redis-utils/pom.xml
@@ -98,6 +98,11 @@
             <artifactId>ballerina-launcher</artifactId>
             <scope>test</scope>
         </dependency> -->
+        <dependency>
+            <groupId>org.ballerinalang</groupId>
+            <artifactId>ballerina-test-utils</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
